feat: refactor e2e test into more reusable functions
